# openapi/openapi.yaml
openapi: 3.0.3
info:
  version: "1.0"
  title: ISURIDE API Specification
servers:
  - url: "http://localhost:8080"
    description: api
paths:
  /initialize:
    post:
      tags:
        - system
      summary: サービスを初期化する
      description: ""
      operationId: post-initialize
      responses:
        "200":
          description: サービスの初期化が完了した
          content:
            application/json:
              schema:
                type: object
                properties:
                  language:
                    type: string
                    enum:
                      - go
                      - perl
                      - php
                      - python
                      - ruby
                      - rust
                      - node
                    description: 実装言語
                required:
                  - language
  /app/users:
    post:
      tags:
        - app
      summary: ユーザーが会員登録を行う
      operationId: app-post-users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: ユーザー名
                firstname:
                  type: string
                  description: 名前
                lastname:
                  type: string
                  description: 名字
                date_of_birth:
                  type: string
                  description: 生年月日
              required:
                - username
                - firstname
                - lastname
                - date_of_birth
      responses:
        "200":
          description: ユーザー登録が完了した
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: アクセストークン
                  id:
                    type: string
                    description: ユーザーID
                required:
                  - access_token
                  - id
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /app/payment-methods:
    post:
      tags:
        - app
      summary: 決済トークンの登録
      operationId: app-post-payment-methods
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: 決済トークン
              required:
                - token
      responses:
        "204":
          description: 決済トークンの登録に成功した
  /app/rides:
    post:
      tags:
        - app
      summary: ユーザーが配車要求を行う
      operationId: app-post-rides
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pickup_coordinate:
                  $ref: "#/components/schemas/Coordinate"
                  description: 配車位置
                destination_coordinate:
                  $ref: "#/components/schemas/Coordinate"
                  description: 目的地
              required:
                - pickup_coordinate
                - destination_coordinate
      responses:
        "202":
          description: 配車要求を受け付けた
          content:
            application/json:
              schema:
                type: object
                properties:
                  ride_id:
                    type: string
                    description: 配車要求ID
                required:
                  - ride_id
  "/app/rides/{ride_id}":
    get:
      tags:
        - app
      summary: ユーザーが配車要求の状態を確認する
      operationId: app-get-ride
      parameters:
        - $ref: "#/components/parameters/ride_id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideForApp"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/app/rides/{ride_id}/evaluation":
    post:
      tags:
        - app
      summary: ユーザーが椅子を評価する
      operationId: app-post-ride-evaluation
      parameters:
        - name: ride_id
          in: path
          description: ライドID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                evaluation:
                  type: integer
                  description: 椅子の評価
                  minimum: 1
                  maximum: 5
              required:
                - evaluation
      responses:
        "204":
          description: ユーザーが椅子を評価した
        "400":
          description: 椅子が目的地に到着していない、ユーザーが乗車していない、すでに到着しているなど
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 存在しない配車要求
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /app/notification:
    get:
      tags:
        - app
      summary: ユーザー向け通知エンドポイント
      description: 最新の自分の配車要求を取得します。
      operationId: app-get-notification
      responses:
        "200":
          description: Server Sent Eventsで通知レスポンスが送信されます
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideForApp"
        "204":
          description: 対象となる配車要求が存在しない場合
  /provider/providers:
    post:
      tags:
        - provider
      summary: 椅子プロバイダーが登録を行う
      operationId: provider-post-providers
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: プロバイダー名
              required:
                - name
      responses:
        "201":
          description: 椅子プロバイダー登録が完了した
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: アクセストークン
                  id:
                    type: string
                    description: プロバイダーID
                required:
                  - access_token
                  - id
  /provider/sales:
    get:
      tags:
        - provider
      summary: 椅子プロバイダーが指定期間の全体・椅子ごと・モデルごとの売上情報を取得する
      operationId: provider-get-sales
      parameters:
        - name: since
          in: query
          description: 開始日（含む）
          required: true
          schema:
            type: string
        - name: until
          in: query
          description: 終了日（含む）
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_sales:
                    type: integer
                    description: プロバイダー全体の売上
                  chairs:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: 椅子ID
                        name:
                          type: string
                          description: 椅子名
                        sales:
                          type: integer
                          description: 椅子ごとの売上
                    description: 椅子ごとの売上情報
                  models:
                    type: array
                    items:
                      type: object
                      properties:
                        model:
                          type: string
                          description: 椅子モデル
                        sales:
                          type: integer
                          description: モデルごとの売上
                    description: モデルごとの売上情報
  /chair/chairs:
    post:
      tags:
        - chair
      summary: 椅子登録を行う
      operationId: chair-post-chairs
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 椅子の名前
                model:
                  type: string
                  description: 椅子のモデル
              required:
                - name
                - model
      responses:
        "201":
          description: 椅子登録が完了した
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: アクセストークン
                  id:
                    type: string
                    description: 椅子ID
                required:
                  - access_token
                  - id
  /chair/activity:
    post:
      tags:
        - chair
      summary: 椅子が配車受付を開始・停止する
      description: ""
      operationId: chair-post-activity
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                is_active:
                  type: boolean
                  description: 配車受付を開始するか停止するか
              required:
                - is_active
      responses:
        "204":
          description: 椅子の配車受付の開始・停止を受理した
  /chair/coordinate:
    post:
      tags:
        - chair
      summary: 椅子が位置情報を送信する
      operationId: chair-post-coordinate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Coordinate"
      responses:
        "204":
          description: サービスが椅子の座標を認識した
  "/chair/rides/{ride_id}":
    get:
      tags:
        - chair
      summary: 椅子がライドの情報を取得する
      description: 椅子向け通知エンドポイントから通知されたidの情報を取得する想定
      operationId: chair-get-rides
      parameters:
        - $ref: "#/components/parameters/ride_id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideForChair"
        "404":
          description: 存在しない配車要求、対象の椅子にマッチングされていない配車要求を取得しようとした場合など
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/chair/rides/{ride_id}/status":
    post:
      tags:
        - chair
      summary: 椅子がライドの状態を更新する
      operationId: chair-post-ride-status
      parameters:
        - $ref: "#/components/parameters/ride_id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - MATCHING
                    - CARRYING
                    - CANCELED
                  description: |
                    ライドの状態

                    MATCHING: マッチングを拒否し、再度マッチング状態に戻す
                    ENROUTE: マッチングを承認し、乗車位置に向かう
                    CARRYING: ユーザーが乗車し、椅子が目的地に向かう
              required:
                - status
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/chair/rides/{ride_id}/payment":
    post:
      tags:
        - chair
      summary: 支払いを実行する
      operationId: chair-post-ride-payment
      parameters:
        - name: ride_id
          in: path
          description: ライドID
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 支払いに成功した
        "400":
          description: 椅子が目的地に到着していないなど
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 存在しない配車要求IDだった
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /chair/notification:
    get:
      tags:
        - chair
      summary: 椅子向け通知エンドポイント
      description: 椅子に配車要求を通知するなどで使う想定
      operationId: chair-get-notification
      responses:
        "200":
          description: 自分に割り当てられた最新の配車要求
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideForChair"
        "204":
          description: 割り当てられた配車要求が存在しない場合
components:
  parameters:
    ride_id:
      name: ride_id
      in: path
      description: ライドID
      required: true
      schema:
        type: string
  schemas:
    Coordinate:
      type: object
      title: Coordinate
      description: 座標情報
      additionalProperties: false
      properties:
        latitude:
          type: integer
          description: 経度
        longitude:
          type: integer
          description: 緯度
      required:
        - latitude
        - longitude
    RideStatus:
      type: string
      enum:
        - MATCHING
        - ENROUTE
        - PICKUP
        - CARRYING
        - ARRIVED
        - COMPLETED
        - CANCELED
      title: RideStatus
      description: |
        ライドの状態

        MATCHING: サービス上でマッチング処理を行なっていて椅子が確定していない
        ENROUTE: 椅子が確定し、乗車位置に向かっている
        PICKUP: 椅子が乗車位置に到着して、ユーザーの乗車を待機している
        CARRYING: ユーザーが乗車し、椅子が目的地に向かっている
        ARRIVED: 目的地に到着した
        COMPLETED: ユーザーの決済・椅子評価が完了した
        CANCELED: 何らかの理由により途中でキャンセルされた(一定時間待ったが椅子を割り当てられなかった場合などを想定)
    Chair:
      type: object
      title: Chair
      description: 簡易椅子情報
      properties:
        id:
          type: string
          description: 椅子ID
        name:
          type: string
          description: 椅子の名前
        model:
          type: string
          description: 椅子のモデル
      required:
        - id
        - name
        - model
    User:
      type: object
      title: User
      description: 簡易ユーザー情報
      properties:
        id:
          type: string
          description: ユーザーID
        name:
          type: string
          description: ユーザー名
      required:
        - id
        - name
    RideForApp:
      description: App向けライド情報
      type: object
      properties:
        ride_id:
          type: string
          description: ライドID
        pickup_coordinate:
          $ref: "#/components/schemas/Coordinate"
        destination_coordinate:
          $ref: "#/components/schemas/Coordinate"
        status:
          $ref: "#/components/schemas/RideStatus"
        chair:
          $ref: "#/components/schemas/Chair"
        created_at:
          type: number
          description: 配車要求日時
        updated_at:
          type: number
          description: 配車要求更新日時
      required:
        - ride_id
        - pickup_coordinate
        - destination_coordinate
        - status
        - created_at
        - updated_at
    RideForChair:
      description: Chair向けライド情報
      type: object
      properties:
        ride_id:
          type: string
          description: ライドID
        user:
          $ref: "#/components/schemas/User"
        destination_coordinate:
          $ref: "#/components/schemas/Coordinate"
        status:
          $ref: "#/components/schemas/RideStatus"
      required:
        - ride_id
        - user
        - destination_coordinate
    Error:
      type: object
      title: Error
      properties:
        message:
          type: string
      required:
        - message